Move backup actions to budgetSlice
